file database

+++ app/routes/main_routes.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, UploadFile, File, HTTPException
 from tasks.task_manager import task_manager
-from utils.file_utils import main_save_uploaded_files
+from utils.file_utils import extract_zip_with_index
 from tasks.task_processing import distribute_files_to_slaves
 import asyncio
 import os
@@ -10,9 +10,9 @@ import aiofiles
 from pathlib import Path
 import httpx
 import zipfile
+import socket
 
-DB_IP = "192.168.3.12:8000"
-
+DB_IP = "192.168.1.107:8000"
 
 router = APIRouter()
 ######
@@ -30,22 +30,24 @@ async def heartbeat(request: HeartbeatRequest):
     return {"message": "Heartbeat received"}
 ######
 
-
-async def start():
+async def manage_tasks():
     os.makedirs("app/incoming", exist_ok=True)
-    print('HUY')
 
     try:
         async with httpx.AsyncClient() as client:
             response = await client.get(f"http://{DB_IP}/download")
 
+        if response.status_code != 200:
+            print(f"Error downloading file: {response.status_code}")
+            print(response.text)  # Вывод текста ошибки для диагностики
+            raise Exception("Failed to download file")
+
+        # Путь для сохранения ZIP-файла
         zip_path = "app/incoming/temp.zip"
         with open(zip_path, "wb") as zip_file:
             zip_file.write(response.content)
 
-        
-        with zipfile.ZipFile(zip_path, "r") as zip_ref:
-            zip_ref.extractall("app/incoming")
+        await extract_zip_with_index("app/incoming/temp.zip", "app/incoming", task_manager.main_file_index)
 
         os.remove(zip_path)
 
@@ -55,17 +57,34 @@ async def start():
             if 'task' in filename:
                 module_name = os.path.basename(filename.split(".")[0])
                 task_module = importlib.import_module(f'incoming.{module_name}')
-                task_module.cut_jpg(f"app/incoming/data{task_manager.main_file_index}.jpg", r"", task_manager.main_file_index)
+                filepath = f"app/incoming/data{task_manager.main_file_index}.jpg"
+                task_module.cut_jpg(filepath, r"", task_manager.main_file_index)
                 folder_path = Path('app/incoming')
                 for file in folder_path.iterdir():
                     if 'part' in file.name:
                         if str(task_manager.main_file_index) in file.name.split("!")[1][0]:
-                            task_manager.add_file_to_queue(f"app/incoming/{file.name}", filename)           
+                            task_manager.add_file_to_queue(f"app/incoming/{file.name}", f"app/incoming/{filename}")           
         if task_manager.available_hosts:
             asyncio.create_task(distribute_files_to_slaves())
         task_manager.main_file_index += 1
+        if len(task_manager.queue) <= len(task_manager.available_hosts):
+            asyncio.create_task(manage_tasks())
 
     except HTTPException as e:
         if e.status_code == 404:
             print(f"No files available for download. {e}")
         raise
+
+g_port = 0
+
+def set_port(port):
+    global g_port
+    g_port = port
+
+@router.on_event("startup")
+async def start():
+    name = socket.gethostbyname(socket.gethostname())
+    global g_port
+    print(f"{name}:{g_port}")
+    if f"{name}:{g_port}" not in task_manager.available_hosts:
+        await manage_tasks()
